requirements.txt
requests
pandas
numpy
matplotlib
python-telegram-bot
pandas_ta
plotly
python-dotenv
mplfinance